requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cached-property==1.5.1
certifi==2019.3.9
chardet==3.0.4
discord.py==1.1.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
typing-extensions==3.7.2
urllib3==1.24.3
websockets==6.0
yarl==1.3.0

